Converted system calls to Ruby code for better cross-platform compatibility. Refs #199

+++ Vagrantfile
@@ -229,54 +229,86 @@ Vagrant.configure(VAGRANTFILE_API_VERSION) do |config|
         kHost.trigger.before [:up, :provision] do
           info "#{Time.now}: setting up Kubernetes master..."
           info "Setting Kubernetes version #{KUBERNETES_VERSION}"
-          sedInplaceArg = OS.mac? ? " ''" : ""
-          system "cp setup.tmpl temp/setup"
-          system "sed -e 's|__KUBERNETES_VERSION__|#{KUBERNETES_VERSION}|g' -i#{sedInplaceArg} ./temp/setup"
-          system "sed -e 's|__MASTER_IP__|#{MASTER_IP}|g' -i#{sedInplaceArg} ./temp/setup"
+
+          # create setup file
+          setupFile = "#{__dir__}/temp/setup"
+
+          # find and replace kubernetes version and master IP in setup file
+          setupFileData = File.read("setup.tmpl")
+          setupFileData = setupFileData.gsub("__KUBERNETES_VERSION__", KUBERNETES_VERSION);
+          setupFileData = setupFileData.gsub("__MASTER_IP__", MASTER_IP);
+
           if enable_proxy
-            system "sed -e 's|__PROXY_LINE__||g' -i#{sedInplaceArg} ./temp/setup"
-            system "sed -e 's|__NO_PROXY__|#{NO_PROXY}|g' -i#{sedInplaceArg} ./temp/setup"
+            # remove __PROXY_LINE__ flag and set __NO_PROXY__
+            setupFileData = setupFileData.gsub("__PROXY_LINE__", "");
+            setupFileData = setupFileData.gsub("__NO_PROXY__", NO_PROXY);
           else
-            system "sed -e '/__PROXY_LINE__/d' -i#{sedInplaceArg} ./temp/setup"
+            # remove lines that start with __PROXY_LINE__
+            setupFileData = setupFileData.gsub(/^\s*__PROXY_LINE__.*$/, "");
           end
+
+          # write new setup data to setup file
+          File.open(setupFile, "w") do |f|
+            f.write(setupFileData)
+          end
+
+          # give setup file executable permissions
           system "chmod +x temp/setup"
 
           info "Downloading Kubernetes binaries if version mismatch or binaries not present.."
           REQUIRED_BINARIES.each do |filename|
-          system("
-            to_download=0
-            file='#{binaries_host_dir}/#{filename}'
-            echo \"Checking for ${file} with version v#{KUBERNETES_VERSION}..\"
+            toDownload = 0
+            file="#{binaries_host_dir}/#{filename}"
 
-            if [ -f \"#{version_file}\" ]; then
-              LAST_VERSION=`cat \"#{version_file}\"`
-            fi
-            if [ \"$LAST_VERSION\" != \"#{KUBERNETES_VERSION}\" ]; then
-              echo \"Versions mismatch [current: $LAST_VERSION, desired: #{KUBERNETES_VERSION}]\"
-              to_download=1
+            info "Checking for #{file} with version v#{KUBERNETES_VERSION}.."
+            if File.exist?("#{version_file}")
+              lastVersion = File.read("#{version_file}").chomp
+            end
+
+            if lastVersion != KUBERNETES_VERSION
+              info "Versions mismatch [current: #{lastVersion}, desired: #{KUBERNETES_VERSION}]"
+              toDownload = 1
             else
-              echo \"Versions match, checking if binary exist..\"
-              if [ ! -f $file ]; then
-                to_download=1
-              fi
-            fi
+              info "Versions match, checking if binary exists..."
+              if !File.exist?(file)
+                toDownload = 1
+              end
+            end
 
-            if [ $to_download == 1 ]; then
-              url=\"https://storage.googleapis.com/kubernetes-release/release/v#{KUBERNETES_VERSION}/bin/linux/amd64/#{filename}\"
-              echo \"Trying to download ${url}..\"
-              (curl -s -k -L -o $file \"$url\") || true
-            fi
-          ")
+            if toDownload == 1
+              urlDomain = "https://storage.googleapis.com"
+              urlResource = "/kubernetes-release/release/v#{KUBERNETES_VERSION}/bin/linux/amd64/#{filename}"
+              info "Trying to download #{urlDomain}#{urlResource}..."
+              Net::HTTP.start(urlDomain) do |http|
+                  resp = http.get(urlResource)
+                  open(file, "wb") do |f|
+                    f.write(resp.body)
+                  end
+              end
+              info "Download complete."
+            end
           end
+
           # only write current version after all files have been downloaded
-          system("echo #{KUBERNETES_VERSION} > #{version_file}")
+          open("#{version_file}", "wb") do |f|
+            f.write("#{KUBERNETES_VERSION}")
+          end
 
           info "Configuring Kubernetes cluster DNS..."
           kHost.vm.provision :file, :source => File.join(File.dirname(__FILE__), "kube-system.yaml"), :destination => "/home/core/kube-system.yaml"
-          system "cp plugins/dns/dns-controller.yaml.tmpl temp/dns-controller.yaml"
-          system "sed -e 's|__MASTER_IP__|#{MASTER_IP}|g' -i#{sedInplaceArg} ./temp/dns-controller.yaml"
-          system "sed -e 's|__DNS_DOMAIN__|#{DNS_DOMAIN}|g' -i#{sedInplaceArg} ./temp/dns-controller.yaml"
-          system "sed -e 's|__DNS_UPSTREAM_SERVERS__|#{DNS_UPSTREAM_SERVERS}|g' -i#{sedInplaceArg} ./temp/dns-controller.yaml"
+
+          # create dns-controller.yaml file
+          dnsControllerFile = "#{__dir__}/temp/dns-controller.yaml"
+          dnsControllerData = File.read("#{__dir__}/plugins/dns/dns-controller.yaml.tmpl")
+
+          dnsControllerData = dnsControllerData.gsub("__MASTER_IP__", MASTER_IP);
+          dnsControllerData = dnsControllerData.gsub("__DNS_DOMAIN__", DNS_DOMAIN);
+          dnsControllerData = dnsControllerData.gsub("__DNS_UPSTREAM_SERVERS__", DNS_UPSTREAM_SERVERS);
+
+          # write new setup data to setup file
+          File.open(dnsControllerFile, "w") do |f|
+            f.write(dnsControllerData)
+          end
 
           if OS.windows?
             kHost.vm.provision :file, :source => File.join(File.dirname(__FILE__), "temp/setup"), :destination => "/home/core/kubectlsetup"
@@ -292,9 +324,11 @@ Vagrant.configure(VAGRANTFILE_API_VERSION) do |config|
       end
 
         kHost.trigger.after [:up, :resume] do
-          info "Sanitizing stuff..."
-          system "ssh-add ~/.vagrant.d/insecure_private_key"
-          system "rm -rf ~/.fleetctl/known_hosts"
+          unless OS.windows?
+            info "Sanitizing stuff..."
+            system "ssh-add ~/.vagrant.d/insecure_private_key"
+            system "rm -rf ~/.fleetctl/known_hosts"
+          end
         end
 
         kHost.trigger.after [:up] do
@@ -324,14 +358,14 @@ Vagrant.configure(VAGRANTFILE_API_VERSION) do |config|
 
           # set cluster
           if OS.windows?
-              run_remote "/opt/bin/kubectl config set-cluster local --server=http://#{MASTER_IP}:8080 --insecure-skip-tls-verify=true"
-              run_remote "/opt/bin/kubectl config set-context local --cluster=local --namespace=default"
-              run_remote "/opt/bin/kubectl config use-context local"
-            else
-              system "kubectl config set-cluster local --server=http://#{MASTER_IP}:8080 --insecure-skip-tls-verify=true"
-              system "kubectl config set-context local --cluster=local --namespace=default"
-              system "kubectl config use-context local"
-            end
+            run_remote "/opt/bin/kubectl config set-cluster local --server=http://#{MASTER_IP}:8080 --insecure-skip-tls-verify=true"
+            run_remote "/opt/bin/kubectl config set-context local --cluster=local --namespace=default"
+            run_remote "/opt/bin/kubectl config use-context local"
+          else
+            system "kubectl config set-cluster local --server=http://#{MASTER_IP}:8080 --insecure-skip-tls-verify=true"
+            system "kubectl config set-context local --cluster=local --namespace=default"
+            system "kubectl config use-context local"
+          end
 
           info "Configuring Kubernetes DNS..."
           res, uri.path = nil, '/api/v1/namespaces/kube-system/replicationcontrollers/kube-dns'
@@ -440,9 +474,7 @@ Vagrant.configure(VAGRANTFILE_API_VERSION) do |config|
       end
 
       kHost.trigger.before [:destroy] do
-        system <<-EOT.prepend("\n\n") + "\n"
-          rm -f temp/*
-        EOT
+        FileUtils.rm_rf(Dir.glob("#{__dir__}/temp/*"))
       end
 
       if SERIAL_LOGGING
